+++ src/main/java/controller/BlackjackController.java
@@ -9,10 +9,10 @@ import domain.card.deck.CardDeck;
 import domain.card.deck.CardDeckFactory;
 import domain.participant.BetAmount;
 import domain.participant.Dealer;
-import domain.participant.Name;
-import domain.participant.Names;
 import domain.participant.Participants;
 import domain.participant.Player;
+import domain.participant.name.Name;
+import domain.participant.name.Names;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.function.Supplier;
