+++ src/main/java/domain/participant/Dealer.java
@@ -1,5 +1,7 @@
 package domain.participant;
 
+import domain.participant.name.Name;
+
 public class Dealer extends Participant {
     private static final Name DEFAULT_NAME = new Name("딜러");
     private static final int DEALER_MIN_SCORE_POLICY = 17;
