+++ src/test/java/fixture/ParticipantFixture.java
@@ -5,9 +5,9 @@ import static fixture.CardFixture.카드;
 import domain.card.Denomination;
 import domain.participant.BetAmount;
 import domain.participant.Dealer;
-import domain.participant.Name;
 import domain.participant.Participants;
 import domain.participant.Player;
+import domain.participant.name.Name;
 import java.util.Arrays;
 import java.util.List;
